Chenweinew

See merge request test1/auto-test!194

+++ testCase/ContractTestCase/07_Noti/ws_trade/TestContractNoti_ws_trade_043.py
@@ -61,7 +61,7 @@ class TestContractNoti_ws_trade_043:
                 assert False
             if tradedetail['direction'] == None:
                 assert False
-            if int(tradedetail['price']) != int(self.current_price):
+            if float(tradedetail['price']) != float(self.current_price):
                 assert False
             if tradedetail['quantity'] == None:
                 assert False
